1. Modify the cininstaller installing path from ~/Application Support/OpenVanillia/OVIMGeneric to ~/Application Support/OpenVanilla/UserData/OVIMGeneric to match userdatapath in OVIMGeneric module. 2. Rewrite deprecated file utility functions to replaced version.

+++ Loaders/OSX-IMK/LVInputController.mm
@@ -299,7 +299,7 @@ static id LVICCurrentlyActiveSender = nil;
 		// hide notify message tooltip if there's already one, otherwise do nothing
 	}
 }
-- (unsigned int)recognizedEvents:(id)sender
+- (NSUInteger)recognizedEvents:(id)sender
 {
     return NSKeyDownMask | NSKeyUpMask | NSFlagsChangedMask | NSMouseEnteredMask | NSLeftMouseDownMask | NSLeftMouseDragged;
 }
